# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telegram/webhook:
    post:
      operationId: receiveTelegramWebhook
      tags: [telegram]
      summary: Receive Telegram Bot API updates
      description: Public webhook endpoint authenticated by Telegram's webhook secret header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramWebhookUpdate"
      responses:
        "200":
          description: Update acknowledged immediately
        "401":
          description: Invalid Telegram webhook secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Telegram webhook is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rm_session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TelegramWebhookUpdate:
      type: object
      additionalProperties: true
...